backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.5
tzdata==2025.3
djangorestframework-simplejwt==5.3.1
youtube-transcript-api
youtube-transcript-api
google-genai
python-dotenv
python-dotenv
